# perf: more efficient node ancestor calculation