When testing on the VSCode repo, with this rule:

```yaml
id: id-in-function
language: typescript
rule:
  kind: identifier
  inside:
    kind: function_declaration
    stopBy: end
```

These were the average times for `ast-grep scan --report-style=short > test.txt` with `AMD Ryzen 5 7640U`:

| Before | After |
| - | - |
| 3.9s | 2.3s |

---

Requires https://github.com/ast-grep/tree-sitter-wasm/pull/3

Explanation: Currently we iterate up the tree by getting the node's parent, in order to find ancestors. However this is slow because internally, tree-sitter does not store parent information in nodes. It retrieves parents by repeatedly calling `ts_node_child_with_descendant`, which retrieves the node's immediate child which contains the target. Instead, we can call this function directly so as to only iterate down the chain once, rather than `n` times.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Refactor**
  - Enhanced the logic for hierarchical data navigation to ensure that related items are retrieved accurately and efficiently.
  - Maintained backward compatibility to preserve a seamless experience for users leveraging connected data.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->