@HerringtonDarkholme 2025-04-13
FYI tree-sitter's endiance does not support napi/pyo3 build. 

See 
https://github.com/tree-sitter/tree-sitter/discussions/4224
https://github.com/ast-grep/ast-grep/actions/runs/14431009646