@ribru17 2025-04-13
Ah, that is unfortunate :/ does tree-sitter 0.24 support this? I believe the new method was actually added in that version