@HerringtonDarkholme 2025-04-13
another way is to fix this one https://github.com/tree-sitter/tree-sitter/blob/master/lib/src/portable/endian.h

to make it compiles on more platforms