@ribru17 2025-04-13
That same build command works on my machine, maybe the docker environment does not define e.g. `__linux__`?