@HerringtonDarkholme 2025-04-13
You can take a look at these run

https://github.com/ast-grep/ast-grep/actions/runs/14431095274/job/40465963122

https://github.com/ast-grep/ast-grep/actions/runs/14431094961

https://github.com/ast-grep/ast-grep/actions/runs/14431010713/job/40465778164


Also, I think you can try run the action manually on your fork.

<img width="463" alt="image" src="https://github.com/user-attachments/assets/16f28023-19ff-47f9-9006-8d3be5fba160" />