@HerringtonDarkholme 2025-04-13
to be more precise, tree-sitter native 0.24.7 has this method, but web-tree-sitter does not