@HerringtonDarkholme 2025-04-13
Hi! @ribru17 I'm also working on make the system work! https://github.com/ast-grep/ast-grep/pull/1929

I am blocked by napi-rs. Hope your image can help!