@HerringtonDarkholme 2025-04-14
it compiled
https://github.com/ast-grep/ast-grep/actions/runs/14435108453/job/40475054324