@HerringtonDarkholme 2025-04-15
inside operator does not work in wasm build

https://ast-grep.github.io/playground.html#eyJtb2RlIjoiQ29uZmlnIiwibGFuZyI6ImphdmFzY3JpcHQiLCJxdWVyeSI6ImNvbnNvbGUubG9nKCRNQVRDSCkiLCJyZXdyaXRlIjoibG9nZ2VyLmxvZygkTUFUQ0gpIiwic3RyaWN0bmVzcyI6InNtYXJ0Iiwic2VsZWN0b3IiOiIiLCJjb25maWciOiJydWxlOlxuICBraW5kOiBzdHJpbmdcbiAgaW5zaWRlOlxuICAgIGtpbmQ6IGltcG9ydF9zdGF0ZW1lbnRcbiAgICBzdG9wQnk6IGVuZCIsInNvdXJjZSI6ImltcG9ydCBhLCB7YiwgYywgZH0gZnJvbSBcIi4vZmlsZVwiO1xuaW1wb3J0IGUgZnJvbSBcIi4vb3RoZXJfZmlsZS5qc1wiO1xuaW1wb3J0IFwiLi9mb2xkZXIvXCI7XG5pbXBvcnQge3h9IGZyb20gXCJwYWNrYWdlXCI7XG5pbXBvcnQge3l9IGZyb20gXCJwYWNrYWdlL3dpdGgvcGF0aFwiO1xuXG5pbXBvcnQoXCIuL2R5bmFtaWMxXCIpO1xuaW1wb3J0KFwiLi9keW5hbWljMi5qc1wiKTtcblxubXlfZnVuYyhcIi4vdW5yZWxhdGVkX3BhdGhfc3RyaW5nXCIpXG5cbiJ9