@HerringtonDarkholme 2025-04-15
I reverted the wasm build, but you can reproduce the issue using this snippet

https://github.com/tree-sitter/tree-sitter/tree/master/lib/binding_web

```ts
const { Parser } = require('web-tree-sitter');
await Parser.init()
const { Language } = require('web-tree-sitter');
const JavaScript = await Language.load('/path/to/tree-sitter-javascript.wasm');
parser.setLanguage(JavaScript);
const sourceCode = 'let x = 1; console.log(x);';
const tree = parser.parse(sourceCode);
const root = tree.rootNode
const child = tree.children[0].children[0]
root.childWithDescendant(child) // null
```