@HerringtonDarkholme 2025-04-15
https://github.com/tree-sitter/tree-sitter/issues/4362