@ribru17 2025-04-15
Ack! so frustrating, classic tree-sitter :sweat_smile: thanks for the report. I've put up a PR: https://github.com/tree-sitter/tree-sitter/pull/4363